Revert "Fix/batchtransfer forms structure"
