+++ src/components/Forms/portfolioTransfters/CoinsToTransfer.tsx
@@ -1,8 +1,7 @@
 import { component$ } from "@builder.io/qwik";
 import { FormBadge } from "~/components/FormBadge/FormBadge";
-
+import { type BatchTransferFormStore } from "~/routes/app/portfolio";
 import IconArrowDown from "/public/assets/icons/arrow-down.svg?jsx";
-import { type BatchTransferFormStore } from "~/routes/app/portfolio/interface";
 
 export interface CoinsToTransferProps {
   batchTransferFormStore: BatchTransferFormStore;
@@ -14,6 +13,7 @@ export default component$<CoinsToTransferProps>(
     return (
       <>
         {availableStructures.value.map((structure: any, index: number) => {
+          console.log(structure);
           return (
             <>
               <div
@@ -26,19 +26,8 @@ export default component$<CoinsToTransferProps>(
                 </div>
                 <div class="mr-2 flex flex-col py-2">
                   {structure.structureBalance.map(
-                    (balance: any, index: number) => {
-                      const currentStructure =
-                        batchTransferFormStore.coinsToTransfer.find(
-                          (struct) => struct.name === structure.structure.name,
-                        );
-
-                      const currentCoin = currentStructure!.coins.find(
-                        (item) =>
-                          item.wallet === balance.wallet.name &&
-                          item.symbol === balance.balance.symbol,
-                      );
-
-                      return balance.wallet.isExecutable ? (
+                    (balance: any, index: number) =>
+                      balance.wallet.isExecutable ? (
                         <FormBadge
                           key={index}
                           text={balance.balance.symbol}
@@ -53,18 +42,48 @@ export default component$<CoinsToTransferProps>(
                               id={`${structure.structure.name}${balance.wallet.name}${balance.balance.symbol}`}
                               name={`${structure.structure.name}${balance.wallet.name}${balance.balance.symbol}`}
                               type="checkbox"
-                              checked={currentCoin!.isChecked}
-                              // value={`${structure.structure.name}${balance.balance.symbol}`}
+                              value={`${structure.structure.name}${balance.balance.symbol}`}
                               class="border-gradient custom-border-1 custom-bg-white checked checked:after:border-bg absolute end-4 z-10  h-6 w-6 appearance-none rounded checked:after:absolute checked:after:left-1/2 checked:after:top-1/2 checked:after:h-2.5 checked:after:w-1.5 checked:after:-translate-x-1/2 checked:after:-translate-y-1/2 checked:after:rotate-45 checked:after:border-solid hover:cursor-pointer focus:after:absolute focus:after:z-[1]"
                               onClick$={() => {
-                                currentCoin!.isChecked =
-                                  !currentCoin?.isChecked;
+                                const x =
+                                  batchTransferFormStore.coinsToTransfer.find(
+                                    (item) =>
+                                      item.name === structure.structure.name,
+                                  );
+                                if (x) {
+                                  const y = x.coins.find(
+                                    (item) =>
+                                      item.wallet === balance.wallet.name,
+                                  );
+                                  if (y) {
+                                    if (
+                                      !y.coins.find(
+                                        (coin) =>
+                                          coin.symbol ===
+                                          balance.balance.symbol,
+                                      )
+                                    ) {
+                                      y.coins.push({
+                                        symbol: balance.balance.symbol,
+                                        amount: "0",
+                                      });
+                                    } else {
+                                      const element = y.coins.find(
+                                        (coin) =>
+                                          coin.symbol ===
+                                          balance.balance.symbol,
+                                      )!;
+                                      const indexToRemove =
+                                        y.coins.indexOf(element);
+                                      y.coins.splice(indexToRemove, 1);
+                                    }
+                                  }
+                                }
                               }}
                             />
                           }
                         />
-                      ) : null;
-                    },
+                      ) : null,
                   )}
                 </div>
               </div>
