+++ src/interface/auth/nonce.ts
@@ -7,9 +7,7 @@ export const CreateNonceResult = z.object({
   exp: z.string().datetime(),
   val: z.string(),
 });
-
 export type CreateNonceResult = z.infer<typeof CreateNonceResult>;
-
 export const createNonce = async (db: Surreal, val: string) => {
   const queryResult = (
     await db.query("CREATE ONLY nonce SET val = type::string($val)", { val })
@@ -21,9 +19,7 @@ export const GetNonceResult = z.object({
   id: z.string(),
   val: z.string(),
 });
-
 export type GetNonceResult = z.infer<typeof GetNonceResult>;
-
 export const getNonce = async (db: Surreal, val: string) => {
   const queryResult = (
     await db.query(
@@ -42,9 +38,7 @@ export const DeleteNonceResult = z.object({
   exp: z.string().datetime(),
   val: z.string(),
 });
-
 export type DeleteNonceResult = z.infer<typeof DeleteNonceResult>;
-
 export const deleteNonce = async (db: Surreal, id: string) => {
   return DeleteNonceResult.parse(
     await db.query(
@@ -62,11 +56,9 @@ export const DeleteExpiredNoncesResult = z.object({
   exp: z.string().datetime(),
   val: z.string(),
 });
-
 export type DeleteExpiredNoncesResult = z.infer<
   typeof DeleteExpiredNoncesResult
 >;
-
 export const deleteExpiredNonces = async (db: Surreal) => {
   const queryResult = (
     await db.query("DELETE nonce WHERE exp < time::now() RETURN BEFORE")
