+++ src/routes/app/portfolio/index.tsx
@@ -14,11 +14,25 @@ import {
   useContext,
 } from "@builder.io/qwik";
 import { messagesContext } from "../layout";
-import { Form } from "@builder.io/qwik-city";
-
+import {
+  Form,
+  routeAction$,
+  routeLoader$,
+  server$,
+  z,
+  zod$,
+} from "@builder.io/qwik-city";
+import { connectToDB } from "~/utils/db";
+import jwt, { type JwtPayload } from "jsonwebtoken";
+import {
+  getDBTokenPriceUSD,
+  getResultAddresses,
+  getWalletDetails,
+} from "~/interface/wallets/observedWallets";
+import { type Wallet } from "~/interface/auth/Wallet";
 import { Modal } from "~/components/Modal/Modal";
 import { isValidName } from "~/utils/validators/addWallet";
-
+import { structureExists } from "~/interface/structure/removeStructure";
 import {
   simulateContract,
   writeContract,
@@ -30,28 +44,282 @@ import { getCookie } from "~/utils/refresh";
 import CoinsToTransfer from "~/components/Forms/portfolioTransfters/CoinsToTransfer";
 import CoinsAmounts from "~/components/Forms/portfolioTransfters/CoinsAmounts";
 import Destination from "~/components/Forms/portfolioTransfters/Destination";
+import { convertToFraction } from "../wallets";
 import { NoDataAdded } from "~/components/NoDataAdded/NoDataAdded";
-import {
-  useDeleteStructure,
-  useDeleteToken,
-  useCreateStructure,
-  useObservedWalletBalances,
-  useAvailableStructures,
-} from "./server";
-export {
-  useDeleteStructure,
-  useDeleteToken,
-  useCreateStructure,
-  useObservedWalletBalances,
-  useAvailableStructures,
-} from "./server";
-import { queryTokens } from "~/database/tokens";
-import { convertToFraction } from "~/utils/fractions";
-import {
-  type WalletWithBalance,
-  type BatchTransferFormStore,
-} from "./interface";
 
+type WalletWithBalance = {
+  wallet: {
+    id: string;
+    chainID: number;
+
+    address: string;
+    isExecutable: boolean;
+  };
+  walletName: string;
+  balance: [{ balanceId: string; tokenId: string; tokenSymbol: string }];
+};
+type CoinObject = {
+  symbol: string;
+  amount: string;
+};
+type CoinToApprove = {
+  wallet: string;
+  isExecutable: boolean;
+  address: string;
+  coins: CoinObject[];
+};
+type StructureToApprove = {
+  name: string;
+  coins: CoinToApprove[];
+};
+export interface BatchTransferFormStore {
+  receiverAddress: string;
+  coinsToTransfer: StructureToApprove[];
+}
+export const useDeleteStructure = routeAction$(
+  async (structure, requestEvent) => {
+    const db = await connectToDB(requestEvent.env);
+
+    const cookie = requestEvent.cookie.get("accessToken");
+    if (!cookie) {
+      throw new Error("No cookie found");
+    }
+    if (!(await structureExists(db, structure.id))) {
+      throw new Error("Structure does not exist");
+    }
+
+    await db.delete(structure.id as string);
+
+    return {
+      success: true,
+      structure: { id: structure.id },
+    };
+  },
+  zod$({
+    id: z.string(),
+  }),
+);
+export const useDeleteToken = routeAction$(
+  async (data, requestEvent) => {
+    const db = await connectToDB(requestEvent.env);
+    const cookie = requestEvent.cookie.get("accessToken");
+    if (!cookie) {
+      throw new Error("No cookie found");
+    }
+    if (!(await structureExists(db, data.structureId))) {
+      throw new Error("Structure does not exist");
+    }
+
+    await db.query(`
+    DELETE structure_balance WHERE in=${data.structureId} AND out=${data.balanceId}`);
+
+    const [balanceCount]: any = await db.query(`
+    RETURN COUNT(SELECT id AS num_rows FROM structure_balance WHERE in=${data.structureId})`);
+
+    if (balanceCount === 0) {
+      await db.delete(data.structureId as string);
+    }
+  },
+  zod$({
+    structureId: z.string(),
+    balanceId: z.string(),
+  }),
+);
+
+export const useObservedWalletBalances = routeLoader$(async (requestEvent) => {
+  const db = await connectToDB(requestEvent.env);
+
+  const cookie = requestEvent.cookie.get("accessToken");
+  if (!cookie) {
+    throw new Error("No cookie found");
+  }
+  const { userId } = jwt.decode(cookie.value) as JwtPayload;
+
+  const resultAddresses = await getResultAddresses(db, userId);
+  if (!resultAddresses[0]) {
+    return [];
+  }
+  const walletsWithBalance = [];
+  let balanceDetails = [];
+
+  for (const observedWalletAddress of resultAddresses) {
+    const walletDetails = await getWalletDetails(
+      db,
+      observedWalletAddress.address,
+      userId,
+    );
+    const [balances]: any = await db.query(
+      `SELECT id, value FROM balance WHERE ->(for_wallet WHERE out = '${walletDetails.id}')`,
+    );
+
+    const walletNameResult: any = await db.query(
+      `SELECT VALUE name FROM ${walletDetails.id}<-observes_wallet WHERE in = ${userId}`,
+    );
+
+    for (const balance of balances) {
+      if (balance.value === "0") {
+        continue;
+      }
+      const [tokenId]: any = await db.query(`
+      SELECT VALUE ->for_token.out FROM ${balance.id}`);
+
+      const [tokenDetails]: any = await db.query(`
+      SELECT * FROM ${tokenId[0]}`);
+
+      const walletBalance = {
+        balanceId: balance.id,
+        tokenId: tokenDetails[0].id,
+        tokenSymbol: tokenDetails[0].symbol,
+      };
+      balanceDetails.push(walletBalance);
+    }
+
+    const walletWithBalance = {
+      wallet: walletDetails,
+      walletName: walletNameResult,
+      balance: balanceDetails,
+    };
+    balanceDetails = [];
+
+    walletsWithBalance.push(walletWithBalance);
+  }
+  return walletsWithBalance;
+});
+
+export const useAvailableStructures = routeLoader$(async (requestEvent) => {
+  const db = await connectToDB(requestEvent.env);
+  const cookie = requestEvent.cookie.get("accessToken");
+  if (!cookie) {
+    throw new Error("No cookie found");
+  }
+  const { userId } = jwt.decode(cookie.value) as JwtPayload;
+
+  const [result]: any = await db.query(`
+    SELECT VALUE ->has_structure.out FROM ${userId}`);
+
+  if (!result) throw new Error("No structures available");
+  const createdStructureQueryResult = result[0];
+  const availableStructures: any[] = [];
+
+  for (const createdStructure of createdStructureQueryResult) {
+    const [structure] = await db.select(`${createdStructure}`);
+    const structureTokens: any = [];
+    const [structureBalances]: any = await db.query(`
+      SELECT VALUE ->structure_balance.out
+      FROM ${structure.id}`);
+
+    if (!structureBalances[0].length) {
+      await db.delete(structure.id);
+    } else {
+      for (const balance of structureBalances[0]) {
+        const [walletId]: any = await db.query(`
+        SELECT out
+        FROM for_wallet
+        WHERE in = ${balance}`);
+
+        const [wallet] = await db.select<Wallet>(`${walletId[0].out}`);
+
+        const [[walletNameResult]]: any = await db.query(
+          `SELECT VALUE name FROM ${wallet.id}<-observes_wallet WHERE in = ${userId}`,
+        );
+
+        const [tokenBalance]: string[] = await db.query(`
+        SELECT VALUE value
+        FROM balance
+        WHERE id = ${balance}`);
+
+        const [tokenId]: any = await db.query(`
+        SELECT VALUE ->for_token.out
+        FROM ${balance}`);
+
+        const [token]: any = await db.query(
+          `SELECT *
+         FROM ${tokenId[0]}`,
+        );
+
+        const [tokenValue] = await getDBTokenPriceUSD(db, token[0].address);
+        const tokenWithBalance = {
+          id: token[0].id,
+          name: token[0].name,
+          symbol: token[0].symbol,
+          decimals: token[0].decimals,
+          balance: tokenBalance[0],
+          balanceValueUSD: tokenValue.priceUSD,
+          balanceId: balance,
+        };
+
+        structureTokens.push({
+          wallet: {
+            id: wallet.id,
+            name: walletNameResult,
+            chainId: wallet.chainId,
+            isExecutable: wallet.isExecutable,
+          },
+          balance: tokenWithBalance,
+        });
+      }
+
+      availableStructures.push({
+        structure: {
+          id: structure.id,
+          name: structure.name,
+        },
+        structureBalance: structureTokens,
+      });
+    }
+  }
+
+  return availableStructures;
+});
+export const useCreateStructure = routeAction$(
+  async (data, requestEvent) => {
+    const cookie = requestEvent.cookie.get("accessToken");
+    if (!cookie) {
+      throw new Error("No cookie found");
+    }
+    const { userId } = jwt.decode(cookie.value) as JwtPayload;
+
+    const db = await connectToDB(requestEvent.env);
+
+    const [namesList]: any = await db.query(`
+    SELECT VALUE out.name FROM ${userId}->has_structure`);
+
+    if (namesList.includes(data.name)) {
+      return {
+        success: false,
+        message: "Name already taken",
+      };
+    }
+    const structure = await db.create("structure", {
+      name: data.name,
+    });
+
+    await db.query(`
+    relate only ${userId}-> has_structure -> ${structure[0].id}`);
+
+    for (const balanceId of data.balancesId) {
+      await db.query(`
+      relate only ${structure[0].id}-> structure_balance -> ${balanceId}`);
+    }
+
+    return {
+      success: true,
+      structure: { name: data.name, balances: data.balancesId },
+    };
+  },
+  zod$({
+    name: z.string().min(2, { message: "structure name invalid" }),
+    balancesId: z.array(z.string()),
+  }),
+);
+const queryTokens = server$(async function () {
+  const db = await connectToDB(this.env);
+
+  const [tokens]: any = await db.query(
+    `SELECT decimals, symbol, address FROM token;`,
+  );
+  return tokens;
+});
 export default component$(() => {
   const modalStore = useContext(ModalStoreContext);
   const clickedToken = useStore({ balanceId: "", structureId: "" });
@@ -129,6 +397,7 @@ export default component$(() => {
   );
 
   const handleBatchTransfer = $(async () => {
+    //  watchAccount
     const cookie = getCookie("accessToken");
 
     if (!cookie) throw new Error("No accessToken cookie found");
@@ -145,20 +414,24 @@ export default component$(() => {
       if (modalStore.config) {
         const argsArray = [];
         for (const cStructure of batchTransferFormStore.coinsToTransfer) {
-          for (const cCoin of cStructure.coins) {
-            const chosenToken = tokens.find(
-              (token: any) => token.symbol === cCoin.symbol.toUpperCase(),
-            );
-            const { numerator, denominator } = convertToFraction(cCoin.amount);
-            const calculation =
-              BigInt(numerator * BigInt(10 ** chosenToken.decimals)) /
-              BigInt(denominator);
-            argsArray.push({
-              from: cCoin.address as `0x${string}`,
-              to: batchTransferFormStore.receiverAddress as `0x${string}`,
-              amount: calculation,
-              token: chosenToken.address as `0x${string}`,
-            });
+          for (const cWallet of cStructure.coins) {
+            for (const cCoin of cWallet.coins) {
+              const chosenToken = tokens.find(
+                (token: any) => token.symbol === cCoin.symbol.toUpperCase(),
+              );
+              const { numerator, denominator } = convertToFraction(
+                cCoin.amount,
+              );
+              const calculation =
+                BigInt(numerator * BigInt(10 ** chosenToken.decimals)) /
+                BigInt(denominator);
+              argsArray.push({
+                from: cWallet.address as `0x${string}`,
+                to: batchTransferFormStore.receiverAddress as `0x${string}`,
+                amount: calculation,
+                token: chosenToken.address as `0x${string}`,
+              });
+            }
           }
         }
         const { request } = await simulateContract(modalStore.config, {
@@ -226,11 +499,9 @@ export default component$(() => {
                       const walletAddress = `${observedWalletsWithBalance.value.find((item) => item.wallet.id === wallet.wallet.id)?.wallet.address}`;
                       coins.push({
                         wallet: wallet.wallet.name,
-                        isExecutable: wallet.wallet.isExecutable,
                         address: walletAddress,
-                        symbol: wallet.balance.symbol,
-                        amount: "0",
-                        isChecked: false,
+                        coins: [],
+                        isExecutable: wallet.wallet.isExecutable,
                       });
                     }
                     batchTransferFormStore.coinsToTransfer.push({
@@ -352,11 +623,9 @@ export default component$(() => {
                       const walletAddress = `${observedWalletsWithBalance.value.find((item) => item.walletName === wallet.wallet.name)?.wallet.address}`;
                       coins.push({
                         wallet: wallet.wallet.name,
-                        isExecutable: wallet.wallet.isExecutable,
                         address: walletAddress,
-                        symbol: wallet.balance.symbol,
-                        amount: "0",
-                        isChecked: false,
+                        coins: [],
+                        isExecutable: wallet.wallet.isExecutable,
                       });
                     }
                     batchTransferFormStore.coinsToTransfer.push({
@@ -383,7 +652,6 @@ export default component$(() => {
                 if (stepsCounter.value === 3) {
                   isTransferModalOpen.value = false;
                   stepsCounter.value = 1;
-
                   await handleBatchTransfer();
                 } else {
                   stepsCounter.value = stepsCounter.value + 1;
