--- src/routes/app/portfolio/interface.ts
@@ -1,27 +1,0 @@
-export type WalletWithBalance = {
-  wallet: {
-    id: string;
-    chainID: number;
-
-    address: string;
-    isExecutable: boolean;
-  };
-  walletName: string;
-  balance: [{ balanceId: string; tokenId: string; tokenSymbol: string }];
-};
-export type CoinToApprove = {
-  wallet: string;
-  isExecutable: string;
-  address: string;
-  symbol: string;
-  amount: string;
-  isChecked: boolean;
-};
-export type StructureToApprove = {
-  name: string;
-  coins: CoinToApprove[];
-};
-export interface BatchTransferFormStore {
-  receiverAddress: string;
-  coinsToTransfer: StructureToApprove[];
-}
