--- src/routes/app/portfolio/server/index.ts
@@ -1,10 +1,0 @@
-import { useDeleteStructure } from "./deleteStructureAction";
-export { useDeleteStructure } from "./deleteStructureAction";
-import { useDeleteToken } from "./deleteTokenAction";
-export { useDeleteToken } from "./deleteTokenAction";
-import { useCreateStructure } from "./createStructureAction";
-export { useCreateStructure } from "./createStructureAction";
-import { useObservedWalletBalances } from './observerWalletBalancesLoader';
-export { useObservedWalletBalances } from './observerWalletBalancesLoader';
-import { useAvailableStructures } from './availableStructuresLoader';
-export { useAvailableStructures } from './availableStructuresLoader';
